Move all styles into a separate folder
