+++ App/Components/ArtifactListView.js
@@ -1,38 +1,11 @@
 import React, { Component } from 'react';
 import {
   ListView,
-  StyleSheet,
   Text,
   Image,
   View,
 } from 'react-native';
-
-const styles = StyleSheet.create({
-  container: {
-    marginTop: 60,
-    flex:1
-  },
-  rowContainer: {
-    padding: 10,
-    flexDirection: 'row'
-  },
-  imageContainer: {
-    margin: 3,
-    flex: 2,
-    alignItems: 'stretch'
-  },
-  contentContainer: {
-    margin: 3,
-    flex: 3,
-    justifyContent: 'center'
-  },
-  listImage: {
-    height: 75
-  },
-  listText: {
-    fontSize: 17
-  }
-});
+import { styles } from './Styles/ArtifactListStyle';
 
 class ArtifactListView extends Component {
