+++ App/Components/MainMapView.js
@@ -1,41 +1,11 @@
 import React, { Component } from 'react';
 import {
-  StyleSheet,
   Text,
   View,
   TouchableWithoutFeedback,
   MapView
 } from 'react-native';
-
-const styles = StyleSheet.create({
-  mapContainer: {
-    flex:1
-  },
-  bottomNav: {
-    flex:2,
-    flexDirection: 'row',
-    justifyContent: 'center'
-  },
-  buttonText: {
-    color: '#48BBEC',
-    fontSize: 18,
-    textAlign: 'center'
-  },
-  map: {
-    height:400,
-    flex: 5
-  },
-  bottomNavButton: {
-    flex:1,
-    backgroundColor: '#FFF',
-    justifyContent: 'center',
-    alignItems: 'center',
-    borderWidth: 2,
-    borderColor: '#48BBEC',
-    padding: 10,
-    margin: 2
-  }
-});
+import { styles } from './Styles/MainMapStyle';
 
 class MainMapView extends Component {
