+++ App/Components/SignUpView.js
@@ -1,74 +1,11 @@
 import React, { Component } from 'react';
 import {
-  StyleSheet,
   Text,
   TextInput,
   Image,
   TouchableHighlight
 } from 'react-native';
-
-const styles = StyleSheet.create({
-  mainContainer: {
-    flex: 1,
-    marginTop: 65,
-    flexDirection: 'column',
-    justifyContent: 'center'
-  },
-  title: {
-    fontSize: 60,
-    paddingBottom: 40,
-    textAlign: 'center',
-    fontFamily: 'Bodoni 72 Smallcaps',
-    backgroundColor: 'rgba(0,0,0,0)',
-    color: 'white'
-  },
-  searchInput: {
-    height: 50,
-    padding: 10,
-    textAlign: 'center',
-    marginLeft: 20,
-    marginRight: 20,
-    marginBottom: 15,
-    fontSize: 23,
-    borderWidth: 2,
-    flexDirection: 'row',
-    borderColor: 'white',
-    color: 'white'
-  },
-  buttonText: {
-    fontSize: 18,
-    color: '#000',
-    alignSelf: 'center'
-  },
-  bgImage: { flex: 1,
-             justifyContent: 'center',
-             resizeMode: 'cover'
-  },
-  button: {
-    height: 45,
-    flexDirection: 'row',
-    borderColor: 'white',
-    borderWidth: 1,
-    marginBottom: 10,
-    marginLeft: 20,
-    marginRight: 20,
-    marginTop: 10,
-    backgroundColor: 'white',
-    alignSelf: 'stretch',
-    justifyContent: 'center'
-  },
-  signIn: {
-    color: 'white',
-    padding: 10,
-    textAlign: 'center'
-  },
-  error: {
-    color: '#ef5350',
-    padding: 10,
-    textAlign: 'center',
-    backgroundColor: 'rgba(0,0,0,0)'
-  }
-});
+import { styles } from './Styles/SignUpStyle';
 
 class SignUpView extends Component {
 
@@ -133,21 +70,21 @@ class SignUpView extends Component {
             style={ styles.searchInput }
             value ={ this.state.name }
             onChangeText={ (name) => this.setState({name}) }
-            placeholder='name'/>
+            placeholder="name"/>
         <TextInput
             style={ styles.searchInput }
             value ={ this.state.email }
             onChangeText={ (email) => this.setState({email}) }
-            placeholder='email'/>
+            placeholder="email"/>
         <TextInput
             style={ styles.searchInput }
             secureTextEntry={ true }
             value ={ this.state.password }
-            onChangeText={ (password) => this.setState({password}) } 
-            placeholder='password'/>
+            onChangeText={ (password) => this.setState({password}) }
+            placeholder="password"/>
         <TouchableHighlight
             style={ styles.button }
-            underlayColor='gray'
+            underlayColor="gray"
             onPress={ this._handleSignUp.bind(this) } >
           <Text style={ styles.buttonText }>Sign Up</Text>
         </TouchableHighlight>
