+++ App/Components/Styles/ArtifactListStyle.js
@@ -1,0 +1,30 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    marginTop: 60,
+    flex:1
+  },
+  rowContainer: {
+    padding: 10,
+    flexDirection: 'row'
+  },
+  imageContainer: {
+    margin: 3,
+    flex: 2,
+    alignItems: 'stretch'
+  },
+  contentContainer: {
+    margin: 3,
+    flex: 3,
+    justifyContent: 'center'
+  },
+  listImage: {
+    height: 75
+  },
+  listText: {
+    fontSize: 17
+  }
+});
+
+export { styles };
