+++ App/Components/Styles/IndexStyle.js
@@ -1,0 +1,21 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1
+  },
+  navBar: {
+    backgroundColor: '#F9F9F9'
+  },
+  title: {
+    fontSize: 20,
+    padding: 10
+  },
+  back: {
+    fontSize: 20,
+    color: '#03a9f4',
+    padding: 10
+  }
+});
+
+export { styles };
