+++ index.ios.js
@@ -11,11 +11,11 @@ import { CameraRollView } from './App/Components/CameraRollView';
 import { SubmitImageView } from './App/Components/SubmitImageView';
 import {
   AppRegistry,
-  StyleSheet,
   Navigator,
   Text,
   TouchableHighlight
 } from 'react-native';
+import { styles } from './App/Components/Styles/IndexStyle';
 
 // Initialize Firebase
 import { ENV } from './environment/environment';
@@ -49,7 +49,7 @@ const TITLES = {
 const NavigationBarRouteMapper = {
   Title: (route, navigator, index, navState) => {
     let title = TITLES[route.name];
-    return (<Text style={styles.title}>{title}</Text>); 
+    return (<Text style={styles.title}>{title}</Text>);
   },
   LeftButton: (route, navigator, index, navState) => {
     if (index === 0) {
@@ -68,24 +68,6 @@ const NavigationBarRouteMapper = {
     return null;
   }
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1
-  },
-  navBar: {
-    backgroundColor: '#F9F9F9'
-  },
-  title: {
-    fontSize: 20,
-    padding: 10
-  },
-  back: {
-    fontSize: 20,
-    color: '#03a9f4',
-    padding: 10
-  }
-});
 
 class Quest extends Component {
   constructor(props) {
